+++ system/src/sheets/DishonoredCharacterSheet.mjs
@@ -5,7 +5,7 @@ export default class DishonoredCharacterSheet extends DishonoredBaseActorSheet {
 
 	/** @override */
 	static get defaultOptions() {
-		return mergeObject(super.defaultOptions, {
+		return foundry.utils.mergeObject(super.defaultOptions, {
 			classes: [
 				"dishonored",
 				"sheet",